Handle start angle and end angle for track
Now you specify a start angle and end angle for the track. This is
specified from 0 to 360 degrees (with 0/360 top/center)

Constants used to rotate everything so that 0 is at the top feel a bit
like a kludge, but it appears to work well enough for now and is
reasonably easy to understand.